added eslint, prettier, proptypes & restructured files
