+++ .eslintignore
@@ -1,0 +1,4 @@
+dist/*
+node_modules/*
+./node_modules/*
+react/no-unescaped-entities
