+++ src/App.js
@@ -1,22 +1,12 @@
-import React from "react";
-import { BrowserRouter as Router, Route } from "react-router-dom";
-import Home from "./Home";
-import Login from "./Authentication/Login";
-import SignUp from "./Authentication/SignUp";
-import { AuthProvider } from "./Authentication/Auth";
-import PrivateRoute from "./Authentication/PrivateRoute";
+import React from 'react';
+import Router from './components/Router';
 
 const App = () => {
   return (
-    <AuthProvider>
-      <Router>
-        <div>
-          <PrivateRoute exact path="/" component={Home} />
-          <Route exact path="/login" component={Login} />
-          <Route exact path="/signup" component={SignUp} />
-        </div>
-      </Router>
-    </AuthProvider>
+    <>
+      {/* ThreeJS canvas */}
+      <Router />
+    </>
   );
 };
