+++ src/Authentication/Auth.js
@@ -1,5 +1,6 @@
 import React, { useEffect, useState } from 'react';
 import app from './base';
+import { node } from 'prop-types';
 
 export const AuthContext = React.createContext();
 
@@ -11,9 +12,12 @@ export const AuthProvider = ({ children }) => {
   }, []);
 
   return (
-    <AuthContext.Provider
-    value={{currentUser}}>
+    <AuthContext.Provider value={{ currentUser }}>
       {children}
     </AuthContext.Provider>
-  )
-}
+  );
+};
+
+AuthProvider.propTypes = {
+  children: node.isRequired
+};
