+++ src/components/PrivateRoute.js
@@ -1,6 +1,7 @@
-import React, { useContext } from "react";
-import { Route, Redirect } from "react-router-dom";
-import { AuthContext } from "./Auth";
+import React, { useContext } from 'react';
+import { Route, Redirect } from 'react-router-dom';
+import { AuthContext } from '../authentication/Auth';
+import { func } from 'prop-types';
 
 const PrivateRoute = ({ component: RouteComponent, ...rest }) => {
   const { currentUser } = useContext(AuthContext);
@@ -8,14 +9,18 @@ const PrivateRoute = ({ component: RouteComponent, ...rest }) => {
     <Route
       {...rest}
       render={routeProps =>
-        !!currentUser ? (
+        currentUser ? (
           <RouteComponent {...routeProps} />
         ) : (
-          <Redirect to={"/login"} />
+          <Redirect to={'/login'} />
         )
       }
     />
   );
+};
+
+PrivateRoute.propTypes = {
+  component: func.isRequired
 };
 
 export default PrivateRoute;
