+++ src/index.js
@@ -2,9 +2,8 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import App from './App';
 import * as serviceWorker from './serviceWorker';
-require('dotenv').config()
+require('dotenv').config();
 
 ReactDOM.render(<App />, document.getElementById('root'));
-
 
 serviceWorker.unregister();
